[#64]refactor: client api 에러관련 동작방식 개선 적용, 페이지내의 component들 분리
