+++ client/src/pages/Login/index.tsx
@@ -43,6 +43,7 @@ import {
   SocialButton,
 } from './index.style';
 import { useHistory, useLocation } from '~/core/Router';
+import { ErrorResponse } from '~/lib/api/types';
 
 const LoginPage: FC = () => {
   const { state } = useLocation();
@@ -67,15 +68,16 @@ const LoginPage: FC = () => {
      * @TODO response에 따라서 로그인에 실패했습니다 말고 서버응답에 따라서 다르게 표시하기.
      * 현재 login()의 리턴값 타입추론이 이상함.
      */
-    const res = await login({
+    await login({
       id,
       password: pw,
-    });
-    if (res.statusCode > 400) {
-      alert('회원가입 실패! 유감!');
-      return;
-    }
-    push('/');
+    })
+      .then(() => {
+        push('/');
+      })
+      .catch((err: ErrorResponse) => {
+        alert(err.message);
+      });
   };
 
   const onSocialLogin = () => {
